+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Documentation fixes.
 
 ### Fixed
+- Do not clean full logs folder. Clean each suite logs folder just before executing it. Clean only the log folder of the executed service when it is ran locally as a single service.
 - Pass command arguments just as were defined. Do not execute path.join on them. closes #18
 - Add local services "on.close" handlers just when each service is started. Do not wait until all services are started to add listeners.